Petal backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
openai==2.14.0
httpx==0.27.0
python-dotenv==1.0.1
pydantic==2.9.0
requests==2.32.5
redis==5.0.8
